Enable ccache depend_mode by default
- This should disable "proprocessed_mode" according to
https://ccache.samba.org/manual/latest.html#_the_depend_mode
but for some reason (version? deprecation?) it's not working
on arch linux. Let' keep it here and check why exactly it's
not workint later.

- Additionally, this may be a workaround for some compilation
errors when clang tool changes (and icecc bundle is updated
accordingly), where we get some undefined var/function errors.
Reseting to icecc_bundle to the original one that added
that file into the cache "fixes" (on may only hides) the error.
Not sure, it might be an icecc issue though.

Signed-off-by: Nick Diego Yamane <nickdiego@igalia.com>